OP-21/Bootloader - Change HW ID to 16bits. Should still be compatible with the old version.
git-svn-id: svn://svn.openpilot.org/OpenPilot/trunk@2398 ebee16cc-31ac-478f-84a7-5cbb03baadba

+++ flight/PiOS/inc/pios_opahrs_proto.h
@@ -122,7 +122,7 @@ struct opahrs_msg_v0_rsp_fwdn_data {
 } __attribute__ ((__packed__));
 
 struct opahrs_msg_v0_rsp_versions {
-	uint8_t hw_version;
+	uint16_t hw_version;
 	uint16_t bl_version;
 	uint32_t fw_crc;
 } __attribute__ ((__packed__));
